 requirements.txt
selenium==4.0.0
pytest==7.0.1
allure-pytest==2.9.45
requests==2.27.1
pytest-rerunfailures==10.2